+++ migrations/00000017_publicai/down.sql
@@ -1,0 +1,1 @@
+DROP TABLE IF EXISTS public_ai_sessions;